reorganize example structures
