+++ examples/dynamics_analysis/1d_qif.py
