+++ examples/dynamics_analysis/1d_system.py
