+++ examples/dynamics_analysis/2d_decision_making_model.py
