+++ examples/dynamics_analysis/2d_fitzhugh_nagumo_model.py
