+++ examples/dynamics_analysis/2d_mean_field_QIF.py
