+++ examples/dynamics_analysis/2d_wilson_cowan_model.py
