+++ examples/dynamics_analysis/3d_hindmarsh_rose_model.py
