+++ examples/dynamics_analysis/3d_reduced_trn_model.py
