+++ examples/dynamics_analysis/4d_HH_model.py
