+++ examples/dynamics_analysis/highdim_RNN_Analysis.py
