+++ examples/dynamics_analysis/highdim_gj_coupled_fhn.py
