+++ examples/dynamics_simulation/Bazhenov_1998_thalamus_aug_response.py
