+++ examples/dynamics_simulation/Brette_2007_COBA.py
