+++ examples/dynamics_simulation/Fazli_2022_gj_coupled_bursting_pituitary_cells.py
