+++ examples/dynamics_simulation/JR_1995_jansen_rit_model.py
