+++ examples/dynamics_simulation/Li_2017_unified_thalamus_oscillation_model.py
