+++ examples/dynamics_simulation/Sanda_2021_hippo-tha-cortex-model.py
