+++ examples/dynamics_simulation/Susin_2021_gamma_oscillation_nets.py
