+++ examples/dynamics_simulation/Vreeswijk_1996_EI_net.py
