+++ examples/dynamics_simulation/Wang_2002_decision_making_spiking.py
