+++ examples/dynamics_simulation/Wu_2008_CANN_1D_oscillatory_tracking.py
