+++ examples/dynamics_simulation/Wu_2008_CANN_2D.py
