+++ examples/dynamics_simulation/multi_scale_COBAHH.py
