+++ examples/dynamics_simulation/whole_brain_simulation_with_fhn.py
