+++ examples/dynamics_training/Bellec_2020_eprop_evidence_accumulation.py
