+++ examples/dynamics_training/Gauthier_2021_ngrc_lorenz.py
@@ -5,7 +5,7 @@
 - Gauthier, D.J., Bollt, E., Griffith, A. et al. Next generation reservoir
   computing. Nat Commun 12, 5564 (2021). https://doi.org/10.1038/s41467-021-25801-2
 
-The main task is forecasting the double-scroll system.
+The main task is forecasting the Lorenz63 strange attractor.
 """
 
 
@@ -14,16 +14,17 @@ import numpy as np
 
 import brainpy as bp
 import brainpy.math as bm
+import brainpy_datasets as bd
 
 bm.enable_x64()
 
 
 def get_subset(data, start, end):
-  res = {'x': data['x'][start: end],
-         'y': data['y'][start: end],
-         'z': data['z'][start: end]}
+  res = {'x': data.xs[start: end],
+         'y': data.ys[start: end],
+         'z': data.zs[start: end]}
   res = bm.hstack([res['x'], res['y'], res['z']])
-  return res.reshape((1,) + res.shape)
+  return res.reshape((1, ) + res.shape)
 
 
 def plot_weights(Wout, coefs, bias=None):
@@ -58,7 +59,7 @@ def plot_weights(Wout, coefs, bias=None):
   plt.show()
 
 
-def plot_double_scroll(ground_truth, predictions):
+def plot_lorenz(ground_truth, predictions):
   fig = plt.figure(figsize=(15, 10))
   ax = fig.add_subplot(121, projection='3d')
   ax.set_title("Generated attractor")
@@ -75,57 +76,62 @@ def plot_double_scroll(ground_truth, predictions):
   plt.show()
 
 
-dt = 0.02
-t_warmup = 10.  # ms
-t_train = 100.  # ms
-t_test = 800.  # ms
+dt = 0.01
+t_warmup = 5.  # ms
+t_train = 10.  # ms
+t_test = 120.  # ms
 num_warmup = int(t_warmup / dt)  # warm up NVAR
 num_train = int(t_train / dt)
 num_test = int(t_test / dt)
 
 # Datasets #
 # -------- #
-data_series = bp.datasets.double_scroll_series(t_warmup + t_train + t_test, dt=dt)
+lorenz_series = bd.chaos.LorenzEq(t_warmup + t_train + t_test,
+                                  dt=dt,
+                                  inits={'x': 17.67715816276679,
+                                         'y': 12.931379185960404,
+                                         'z': 43.91404334248268})
 
-X_warmup = get_subset(data_series, 0, num_warmup - 1)
-Y_warmup = get_subset(data_series, 1, num_warmup)
-X_train = get_subset(data_series, num_warmup - 1, num_warmup + num_train - 1)
+X_warmup = get_subset(lorenz_series, 0, num_warmup - 1)
+Y_warmup = get_subset(lorenz_series, 1, num_warmup)
+X_train = get_subset(lorenz_series, num_warmup - 1, num_warmup + num_train - 1)
 # Target: Lorenz[t] - Lorenz[t - 1]
-dX_train = get_subset(data_series, num_warmup, num_warmup + num_train) - X_train
-X_test = get_subset(data_series,
+dX_train = get_subset(lorenz_series, num_warmup, num_warmup + num_train) - X_train
+X_test = get_subset(lorenz_series,
                     num_warmup + num_train - 1,
                     num_warmup + num_train + num_test - 1)
-Y_test = get_subset(data_series,
+Y_test = get_subset(lorenz_series,
                     num_warmup + num_train,
                     num_warmup + num_train + num_test)
 
+
 # Model #
 # ----- #
-
-
 class NGRC(bp.dyn.DynamicalSystem):
   def __init__(self, num_in):
     super(NGRC, self).__init__()
-    self.r = bp.layers.NVAR(num_in, delay=2, order=3, mode=bp.modes.batching)
-    self.di = bp.layers.Dense(self.r.num_out, num_in, mode=bp.modes.training)
+    self.r = bp.layers.NVAR(num_in, delay=2, order=2, constant=True, mode=bp.modes.batching)
+    self.di = bp.layers.Dense(self.r.num_out, num_in, b_initializer=None, mode=bp.modes.training)
 
   def update(self, sha, x):
-    di = self.di(sha, self.r(sha, x))
-    return x + di
+    dx = self.di(sha, self.r(sha, x))
+    return x + dx
 
 
 model = NGRC(3)
+print(model.r.get_feature_names(for_plot=True))
+
 
 # Training #
 # -------- #
 
 # warm-up
-trainer = bp.train.RidgeTrainer(model, alpha=1e-5, jit=True)
+trainer = bp.train.RidgeTrainer(model)
 outputs = trainer.predict(X_warmup)
 print('Warmup NMS: ', bp.losses.mean_squared_error(outputs, Y_warmup))
 
 # training
-trainer.fit([X_train, {'di': dX_train}])
+trainer.fit([X_train, dX_train])
 plot_weights(model.di.W, model.r.get_feature_names(for_plot=True), model.di.b)
 
 # prediction
@@ -134,6 +140,6 @@ model_jit = bm.jit(model)
 outputs = [model_jit(shared, X_test[:, 0])]
 for i in range(1, X_test.shape[1]):
   outputs.append(model_jit(shared, outputs[i - 1]))
-outputs = bm.asarray(outputs).squeeze()
+outputs = bm.asarray(outputs)
 print('Prediction NMS: ', bp.losses.mean_squared_error(outputs, Y_test))
-plot_double_scroll(bm.as_numpy(Y_test).squeeze(), bm.as_numpy(outputs))
+plot_lorenz(bm.as_numpy(Y_test).squeeze(), bm.as_numpy(outputs).squeeze())
