+++ examples/dynamics_training/Gauthier_2021_ngrc_lorenz_inference.py
@@ -8,6 +8,7 @@
 The main task is forecasting the Lorenz63 strange attractor.
 """
 
+import brainpy_datasets as bp_data
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -18,9 +19,9 @@ bm.enable_x64()
 
 
 def get_subset(data, start, end):
-  res = {'x': data['x'][start: end],
-         'y': data['y'][start: end],
-         'z': data['z'][start: end]}
+  res = {'x': data.xs[start: end],
+         'y': data.ys[start: end],
+         'z': data.zs[start: end]}
   X = bm.hstack([res['x'], res['y']])
   X = X.reshape((1,) + X.shape)
   Y = res['z']
@@ -128,11 +129,11 @@ num_test = int(t_test / dt)
 
 # Datasets #
 # -------- #
-lorenz_series = bp.datasets.lorenz_series(t_warmup + t_train + t_test,
-                                          dt=dt,
-                                          inits={'x': 17.67715816276679,
-                                                 'y': 12.931379185960404,
-                                                 'z': 43.91404334248268})
+lorenz_series = bp_data.chaos.LorenzEq(t_warmup + t_train + t_test,
+                                       dt=dt,
+                                       inits={'x': 17.67715816276679,
+                                              'y': 12.931379185960404,
+                                              'z': 43.91404334248268})
 
 X_warmup, Y_warmup = get_subset(lorenz_series, 0, num_warmup)
 X_train, Y_train = get_subset(lorenz_series, num_warmup, num_warmup + num_train)
@@ -170,7 +171,7 @@ trainer.fit([X_train, Y_train])
 outputs = trainer.predict(X_test, reset_state=True)
 print('Prediction NMS: ', bp.losses.mean_squared_error(outputs, Y_test))
 
-plot_lorenz(x=bm.as_numpy(lorenz_series['x']).flatten(),
-            y=bm.as_numpy(lorenz_series['y']).flatten(),
-            true_z=bm.as_numpy(lorenz_series['z']).flatten(),
+plot_lorenz(x=bm.as_numpy(lorenz_series.xs).flatten(),
+            y=bm.as_numpy(lorenz_series.ys).flatten(),
+            true_z=bm.as_numpy(lorenz_series.zs).flatten(),
             predict_z=bm.as_numpy(outputs).flatten())
