+++ examples/dynamics_training/Song_2016_EI_RNN.py
