+++ examples/dynamics_training/SurrogateGrad_lif-ANN-style.py
