+++ examples/dynamics_training/SurrogateGrad_lif_fashion_mnist.py
